validaciones con JWT

+++ controllers/api/apiProductsController.js
@@ -1,9 +1,11 @@
 import Product from "../../models/Products.js";
 import { unlink } from "node:fs/promises";
 import path from "node:path";
+import createError from "http-errors";
 
 export async function list(req, res, next) {
   try {
+    const userId = req.apiUserId;
     const filterName = req.query.name;
     const filterPrice = req.query.price;
     const limit = req.query.limit;
@@ -12,8 +14,7 @@ export async function list(req, res, next) {
     const fields = req.query.fields;
 
     const filter = {
-      // TODO Implement API authentication
-      // owner: userId,
+      owner: userId,
     };
 
     if (filterName) {
@@ -34,7 +35,8 @@ export async function list(req, res, next) {
 export async function getOne(req, res, next) {
   try {
     const productId = req.params.productId;
-    const product = await Product.findById(productId);
+    const userId = req.apiUserId;
+    const product = await Product.findOne({ _id: productId, owner: userId });
 
     res.json({ result: product });
   } catch (error) {
@@ -45,8 +47,10 @@ export async function getOne(req, res, next) {
 export async function newProduct(req, res, next) {
   try {
     const productData = req.body;
+    const userId = req.apiUserId;
     const product = new Product(productData);
     product.avatar = req.file?.filename;
+    product.owner = userId;
     const savedProduct = await product.save();
 
     res.status(201).json({ result: savedProduct });
@@ -58,10 +62,14 @@ export async function newProduct(req, res, next) {
 export async function update(req, res, next) {
   try {
     const productId = req.params.productId;
+    const userId = req.apiUserId;
     const productData = req.body;
     productData.avatar = req.file?.filename;
-    const updatedProduct = await Product.findByIdAndUpdate(
-      productId,
+    const updatedProduct = await Product.findOneAndUpdate(
+      {
+        _id: productId,
+        owner: userId,
+      },
       productData,
       {
         new: true,
@@ -77,7 +85,29 @@ export async function update(req, res, next) {
 export async function deleteProduct(req, res, next) {
   try {
     const productId = req.params.productId;
+    const userId = req.apiUserId;
+
+    // validar que el producto que queremos borrar pertenece al usuario
     const product = await Product.findById(productId);
+
+    // comprobar que existe
+    if (!product) {
+      console.log(
+        `WARINING!!! user ${userId} is trying to delete non existing agent`
+      );
+      next(createError(404));
+      return;
+    }
+
+    //comprobar la propiedad
+    if (product.owner.toString() !== userId) {
+      console.log(
+        `WARINING!!! user ${userId} is trying to delete products of other users`
+      );
+      next(createError(401));
+      return;
+    }
+
     if (product.avatar) {
       await unlink(
         path.join(
